add histogram coloring

+++ inc/fractal/fractal.h
@@ -2,5 +2,5 @@
 
 class Fractal {
   public:
-    virtual double Calculate(const double& x, const double& y) = 0;
+    virtual int Calculate(const double& x, const double& y) = 0;
 };
